Replace InstancesDao with InstancesRepository

+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/InstanceDeleter.java
@@ -19,13 +19,25 @@ public class InstanceDeleter {
 
     public void delete(Long id) {
         Instance instance = instancesRepository.get(id);
-        instancesRepository.delete(id);
+        if (instance != null) {
+            if (instance.getStatus().equals(Instance.STATUS_SUBMITTED)) {
+                instancesRepository.save(new Instance.Builder(instance)
+                        .geometry(null)
+                        .geometryType(null)
+                        .build()
+                );
+                instancesRepository.softDelete(id);
+            } else {
+                instancesRepository.delete(id);
+            }
 
-        Form form = formsRepository.getLatestByFormIdAndVersion(instance.getJrFormId(), instance.getJrVersion());
-        if (form != null && form.isDeleted()) {
-            List<Instance> otherInstances = instancesRepository.getAllNotDeletedByFormIdAndVersion(form.getJrFormId(), form.getJrVersion());
-            if (otherInstances.isEmpty()) {
-                formsRepository.delete(form.getId());
+
+            Form form = formsRepository.getLatestByFormIdAndVersion(instance.getJrFormId(), instance.getJrVersion());
+            if (form != null && form.isDeleted()) {
+                List<Instance> otherInstances = instancesRepository.getAllNotDeletedByFormIdAndVersion(form.getJrFormId(), form.getJrVersion());
+                if (otherInstances.isEmpty()) {
+                    formsRepository.delete(form.getId());
+                }
             }
         }
     }
